requirements.txt
# altair==4.1.0
# streamlit==1.10.0
# google-auth==2.6.6
# google-auth-oauthlib==0.4.6
# google-auth-httplib2==0.1.0
# google-api-python-client==2.37.0
# PyPDF2==1.26.0
# pytesseract==0.3.8
# python-dotenv
# openai==0.13.0
# firebase-admin==5.1.0
# spacy==3.2.4
# nltk==3.7
streamlit
python-dotenv
pytesseract
openai
google-auth-oauthlib
google-api-python-client
PyPDF2
firebase-admin
spacy==3.2.0
nltk
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
